// src/sdk/generated/models/BadgeUpdate.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type BadgeUpdate = {
  display_name?: (string | null);
  description?: (string | null);
  icon_url?: (string | null);
  color_code?: (string | null);
  is_active?: (boolean | null);
  criteria?: (Record<string, any> | null);
};